requirements.txt
mysql-connector-python
birdnet_analyzer
schedule
